run k8s integration tests
